LayoutTests/fast/loader/page-dismissal-modal-dialogs-expected.txt

frame "<!--frame1-->" - has 1 onunload handler(s)
CONSOLE MESSAGE: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.confirm is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.prompt is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.confirm is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.prompt is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.confirm is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.prompt is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.confirm is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.prompt is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.confirm is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.prompt is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.confirm is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.prompt is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.confirm is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.prompt is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.confirm is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.prompt is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.confirm is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.prompt is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.confirm is not allowed while unloading a page.
CONSOLE MESSAGE: Use of window.prompt is not allowed while unloading a page.
PASS
